Create GradeLetter class
A new class is created in order to keep all the Grade Letters in it and
by adding the proper functions and properties in this class, access is
given to the rest part of the application to use this class whenever
needed. The grade letters were used as an array of string in the system
and they were scattered in different places in the project. By having
this separate class we’ll increase the reuse, modifiability and
maintenance.