Extract Method checkingClassNames
The method SaveClass is a kinda big method having almost 90 lines of
code. To make this method cleaner and easier to read, understand and
modify the related lines are extracted to a new method and are called in
main method (SaveClass).
CheckingClassNames method is taking care of checking the Class Criteria
and it returns a bool by validating the different conditions.